Move the existence check for block hash in database operations
